move duplicated code to a separate class
