+++ src/main/java/io/jenkins/plugins/extratoolinstallers/installers/IsAlreadyOnPath.java
@@ -1,13 +1,10 @@
 package io.jenkins.plugins.extratoolinstallers.installers;
 
-import com.google.common.base.Joiner;
 import hudson.Extension;
 import hudson.FilePath;
-import hudson.Launcher;
 import hudson.Util;
 import hudson.model.Node;
 import hudson.model.TaskListener;
-import hudson.tools.ToolInstallation;
 import hudson.tools.ToolInstaller;
 import hudson.tools.ToolInstallerDescriptor;
 import hudson.util.FormValidation;
@@ -22,43 +19,20 @@ import org.kohsuke.stapler.QueryParameter;
 import javax.annotation.CheckForNull;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.io.OutputStream;
-import java.nio.charset.StandardCharsets;
-import java.util.Arrays;
 import java.util.regex.Pattern;
 import java.util.regex.PatternSyntaxException;
 
 /**
  * A {@link ToolInstaller} that locates an existing tool on the agent, or fails.
  */
-public class IsAlreadyOnPath extends ToolInstaller {
+public class IsAlreadyOnPath extends VersionCheckingToolInstaller {
     @CheckForNull
     private String executableName;
 
-    @CheckForNull
-    private String relativePath;
-
-    @CheckForNull
-    private String[] versionCmd;
-
-    @CheckForNull
-    private Pattern versionPattern;
-
-    /** Only used if we've been given a pattern that can't be compiled */
-    @CheckForNull
-    private String versionPatternString;
-
-    @CheckForNull
-    private String versionMin;
-
-    @CheckForNull
-    private String versionMax;
-
     /**
      * Constructor that sets mandatory fields.
-     * 
+     *
      * @param label The {@link ToolInstaller#getLabel()}.
      */
     @DataBoundConstructor
@@ -68,7 +42,7 @@ public class IsAlreadyOnPath extends ToolInstaller {
 
     /**
      * Name of the executable we are to locate.
-     * 
+     *
      * @return Name, or null if none has been set.
      */
     @CheckForNull
@@ -78,7 +52,7 @@ public class IsAlreadyOnPath extends ToolInstaller {
 
     /**
      * Sets {@link #getExecutableName()}.
-     * 
+     *
      * @param executable New value.
      */
     @DataBoundSetter
@@ -86,226 +60,15 @@ public class IsAlreadyOnPath extends ToolInstaller {
         this.executableName = Util.fixEmpty(executable);
     }
 
-    /**
-     * Directory of the tool's "home", relative to wherever we found the command on
-     * the path.
-     * 
-     * @return The relative path, or null if none has been set (which means ".").
-     */
-    @CheckForNull
-    public String getRelativePath() {
-        return Util.fixEmpty(relativePath);
-    }
-
-    /**
-     * Sets {@link #getRelativePath()}.
-     * 
-     * @param relativePath New value.
-     */
-    @DataBoundSetter
-    public void setRelativePath(@Nullable String relativePath) {
-        this.relativePath = Util.fixEmpty(relativePath);
-    }
-
-    /**
-     * Command we run in order to test what version we've got.
-     * 
-     * @return null if not set, else a list where the first element is the command
-     *         and the remaining elements are arguments for the command.
-     */
-    @CheckForNull
-    public String[] getVersionCmd() {
-        return fixEmpty(versionCmd);
-    }
-
-    /**
-     * See {@link #getVersionCmd()}.
-     * 
-     * @return {@link #getVersionCmd()} as a multi-line string.
-     */
     @Nonnull
-    public String getVersionCmdString() {
-        final String[] v = getVersionCmd();
-        if (v == null) {
-            return "";
-        }
-        return Joiner.on('\n').join(v);
-    }
-
-    /**
-     * Sets {@link #getVersionCmd()}.
-     * 
-     * @param versionCmd New value.
-     */
-    public void setVersionCmd(String[] versionCmd) {
-        if (versionCmd != null) {
-            this.versionCmd = Arrays.copyOf(versionCmd, versionCmd.length);
-        } else {
-            this.versionCmd = null;
-        }
-    }
-
-    /**
-     * See {@link #setVersionCmd(String[])}.
-     * 
-     * @param versionCmdString New value as multi-line string.
-     */
-    @DataBoundSetter
-    public void setVersionCmdString(String versionCmdString) {
-        setVersionCmd(Util.fixNull(versionCmdString).split("\n"));
-    }
-
-    private static String[] fixEmpty(String[] l) {
-        if (l == null || l.length==0) {
-            return null;
-        }
-        return l;
-    }
-
-    /**
-     * The regular expression used to parse the output from running
-     * {@link #getVersionCmd()}.
-     * 
-     * @return The regex that was set, or null if not set to a valid value.
-     */
-    public Pattern getVersionPattern() {
-        return versionPattern;
-    }
-
-    /**
-     * Sets {@link #getVersionPatternString()} and {@link #getVersionPattern()}.
-     * 
-     * @param versionPattern New value.
-     */
-    public void setVersionPattern(Pattern versionPattern) {
-        this.versionPattern = versionPattern;
-        this.versionPatternString = null;
-    }
-
-    /**
-     * The regular expression used to parse the output from running
-     * {@link #getVersionCmd()}.
-     * 
-     * @return The regex that was set, or null if not set.
-     */
-    @CheckForNull
-    public String getVersionPatternString() {
-        if( versionPattern!=null ) {
-            return Util.fixEmpty(versionPattern.pattern());
-        }
-        return Util.fixEmpty(versionPatternString);
-    }
-
-    /**
-     * Sets {@link #getVersionPatternString()} and {@link #getVersionPattern()}.
-     * 
-     * @param versionPatternString New value.
-     */
-    @DataBoundSetter
-    public void setVersionPatternString(String versionPatternString) {
-        if (Util.fixEmpty(versionPatternString) != null) {
-            try {
-                this.versionPattern = Pattern.compile(versionPatternString);
-                this.versionPatternString = null;
-            } catch (PatternSyntaxException ex) {
-                this.versionPattern = null;
-                this.versionPatternString = versionPatternString;
-            }
-        } else {
-            this.versionPattern = null;
-            this.versionPatternString = null;
-        }
-    }
-
-    /**
-     * The minimum version acceptable.
-     * 
-     * @return The version that was set, or null if not set.
-     */
-    @CheckForNull
-    public String getVersionMin() {
-        return versionMin;
-    }
-
-    /**
-     * Sets {@link #getVersionMin()}.
-     * 
-     * @param versionMin New value.
-     */
-    @DataBoundSetter
-    public void setVersionMin(String versionMin) {
-        this.versionMin = versionMin;
-    }
-
-
-    /**
-     * The maximum version acceptable.
-     * 
-     * @return The version that was set, or null if not set.
-     */
-    @CheckForNull
-    public String getVersionMax() {
-        return versionMax;
-    }
-
-    /**
-     * Sets {@link #getVersionMax()}.
-     * 
-     * @param versionMax New value.
-     */
-    @DataBoundSetter
-    public void setVersionMax(String versionMax) {
-        this.versionMax = versionMax;
-    }
-
     @Override
-    public FilePath performInstallation(@Nonnull ToolInstallation tool, @Nonnull Node node,
-            @CheckForNull TaskListener log) throws IOException, InterruptedException {
-        final String exeName = getExecutableName();
-        if (exeName == null) {
-            throw new IllegalArgumentException(Messages.IsAlreadyOnPath_executableNameIsEmpty());
-        }
-        final FilePath executablePath = findExecutableOnNodeOrThrow(exeName, node, log);
-        final FilePath parent = executablePath.getParent();
-        if (parent == null) {
-            // This shouldn't happen, hence not localized.
-            throw new IllegalStateException(
-                    "Executable (" + exeName + ") found at '" + executablePath + "' has no parent folder");
-        }
-        final String relPathOrNull = getRelativePath();
-        final FilePath resultToReturn;
-        if (relPathOrNull == null || relPathOrNull.equals(".")) {
-            resultToReturn = parent;
-        } else {
-            resultToReturn = parent.child(relPathOrNull);
-        }
-        final String[] vCmd = getVersionCmd();
-        final Pattern vPattern = getVersionPattern();
-        final String vMax = getVersionMax();
-        final String vMin = getVersionMin();
-        if (vCmd != null && vPattern != null && (vMin != null || vMax != null)) {
-            final ByteArrayOutputStream output = new ByteArrayOutputStream();
-            final Launcher launcher = node.createLauncher(log);
-            runCommandOnNode(launcher, resultToReturn, vCmd, output);
-            final String cmdOutput = output.toString(StandardCharsets.UTF_8.name());
-            final String parsedVersion = VersionChecker.parseVersionCmdOutputForVersion(vPattern, cmdOutput);
-            final int versionComparisonResult = VersionChecker.checkVersionIsInRange(vMin, vMax, parsedVersion);
-            if (versionComparisonResult != 0) {
-                throw new WrongVersionException(exeName, resultToReturn.getRemote(), parsedVersion, vMin,
-                        vMax);
-            }
-        }
-        return resultToReturn;
-    }
-
-    @Nonnull
-    private FilePath findExecutableOnNodeOrThrow(@Nonnull final String exeName, @Nonnull Node node,
-            @CheckForNull final TaskListener logOrNull) throws IOException, InterruptedException {
+    FilePath findExecutableOnNodeOrThrow(@Nonnull Node node,
+                                         @CheckForNull final TaskListener logOrNull) throws IOException, InterruptedException {
+        final FindOnPathCallable nodeOperation = mkCallable(logOrNull);
         final FilePath rootPath = node.getRootPath();
         if (rootPath == null) {
             throw new IllegalStateException(Messages.IsAlreadyOnPath_agentIsOffline());
         }
-        final FindOnPathCallable nodeOperation = mkCallable(exeName, logOrNull);
         final String absolutePathToExecutable = rootPath.act(nodeOperation);
         final FilePath executablePath = node.createPath(absolutePathToExecutable);
         if (executablePath == null) {
@@ -314,14 +77,16 @@ public class IsAlreadyOnPath extends ToolInstaller {
         return executablePath;
     }
 
-    // package access for test purposes only
-    @Restricted(NoExternalUse.class)
-    void runCommandOnNode(final Launcher launcher, final FilePath pwd, final String[] cmd,
-            final OutputStream output) throws IOException, InterruptedException {
-        launcher.launch().cmds(cmd).stdout(output).pwd(pwd).join();
+    @Nonnull
+    @Override
+    FindOnPathCallable mkCallable(@CheckForNull final TaskListener logOrNull) {
+        final String exeName = getExecutableName();
+        if (exeName == null) {
+            throw new IllegalArgumentException(Messages.IsAlreadyOnPath_executableNameIsEmpty());
+        }
+        return mkCallable(exeName, logOrNull);
     }
 
-    // package access for test purposes only
     @Restricted(NoExternalUse.class)
     @Nonnull
     FindOnPathCallable mkCallable(@Nonnull final String exeName, @CheckForNull final TaskListener logOrNull) {
@@ -353,7 +118,7 @@ public class IsAlreadyOnPath extends ToolInstaller {
         }
 
         public FormValidation doCheckVersionPatternString(@QueryParameter String versionCmdString,
-                @QueryParameter String versionPatternString) {
+                                                          @QueryParameter String versionPatternString) {
             if (Util.fixEmpty(versionCmdString) == null) {
                 return FormValidation.ok();
             }
@@ -370,8 +135,8 @@ public class IsAlreadyOnPath extends ToolInstaller {
         }
 
         public FormValidation doCheckVersionMin(@QueryParameter String versionCmdString,
-                @QueryParameter String versionPatternString, @QueryParameter String versionMin,
-                @QueryParameter String versionMax) {
+                                                @QueryParameter String versionPatternString, @QueryParameter String versionMin,
+                                                @QueryParameter String versionMax) {
             if (Util.fixEmpty(versionCmdString) != null && Util.fixEmpty(versionPatternString) != null) {
                 if (Util.fixEmpty(versionMin) == null && Util.fixEmpty(versionMax) == null) {
                     return FormValidation.error(Messages.IsAlreadyOnPath_versionMinMaxNotSpecified());
@@ -391,8 +156,8 @@ public class IsAlreadyOnPath extends ToolInstaller {
         }
 
         public FormValidation doCheckVersionTestString(@QueryParameter String versionTestString,
-                @QueryParameter String versionCmdString, @QueryParameter String versionPatternString,
-                @QueryParameter String versionMin, @QueryParameter String versionMax) {
+                                                       @QueryParameter String versionCmdString, @QueryParameter String versionPatternString,
+                                                       @QueryParameter String versionMin, @QueryParameter String versionMax) {
             if (Util.fixEmpty(versionTestString) == null) {
                 return FormValidation.ok();
             }
@@ -404,8 +169,8 @@ public class IsAlreadyOnPath extends ToolInstaller {
                 return FormValidation.warning(Messages.IsAlreadyOnPath_noVersionValidation() + "\n"
                         + Messages.IsAlreadyOnPath_versionPatternIsEmpty());
             }
-            if (Util.fixEmpty(versionMin) == null ) {
-                if (Util.fixEmpty(versionMax) == null ) {
+            if (Util.fixEmpty(versionMin) == null) {
+                if (Util.fixEmpty(versionMax) == null) {
                     return FormValidation.warning(Messages.IsAlreadyOnPath_noVersionValidation() + "\n"
                             + Messages.IsAlreadyOnPath_versionMinMaxNotSpecified());
                 }
